円弧: 小さい円のピクセル判定を「角度」から本物の箱に、傾きは16.16で読む
小さい円を点で描く経路 (FUN_20a9_0d1a) は、掃引を角度の対で受け取っては
いなかった。最大五つの「破片」の表 -- 象限ひとつと画面座標の箱ひとつ -- を
受け取り、歩きが出す各点を象限へ折り返して、その箱の中にあるものだけを
描く。箱は軸に平行な長方形なので角度そのものとは違い、終端角を少し過ぎた
点でも箱の中なら描かれる。どんな角度判定でも作れない振る舞いで、これまで
合わせ込もうとしていたのはこれだった。

表は当てずっぽうではなく本物から読んだ:

  DOSEMU_BP=11B9:0D1A DOSEMU_BPPTR=8 DOSEMU_BPPTRAT=40 DOSEMU_BPPTRN=50

中心 (150,120)、半径 5、0..22 度の円弧ひとつで、象限 0 / x 154..155 /
y 118..120 が出る。中心を (150.5,120.25)、半径を 5.6 にしても同じ数字が
出るので、箱は切り捨てた中心と半径から作られている。

あわせて、円弧レコードの傾きは 16.16 の long だった。SAMPLE2 には
0x0059ffeb = 88.99968 度で傾いた楕円が四つあり、上位ワードだけを読むと
その連鎖の頂点がひとつずつずれる。

  総差分 192 -> 154、十四図面のうち十一が完全一致
  (SAMPLE2 20 -> 0、SAMPLE6 16 -> 0)

tools/qarc.py は小円の掃引を、tools/qpoly.py は連鎖の頂点を、本物と並べて
聞くための道具。dosv_emu_cpp 側に DOSEMU_BPPTRAT/DOSEMU_BPPTRN を足した。

+++ src/jwc.h
@@ -35,7 +35,7 @@ typedef struct {
     short flatten;      /* minor/major x10000; 10000 is a circle */
     long start;         /* 16.16 fixed degrees */
     long end;           /* end == start means the whole ellipse */
-    short tilt;         /* degrees the ellipse is turned by */
+    long tilt;          /* how far the ellipse is turned, 16.16 degrees too */
     unsigned char type, pen;
     unsigned char layer;
     unsigned char rest[4];
